requirements.txt
Flask==2.3.3
yt-dlp==2023.7.6
python-dotenv==1.0.0
Flask-Cors==4.0.0
requests==2.31.0
PyJWT==2.8.0
Werkzeug==2.3.7
